pyproject.toml
[project]
name = "life-coach"
version = "0.1.0"
description = "OpenAI SDK, assignment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphviz>=0.21",
    "openai-agents[viz]>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]